Compile EEPROM emulation to device
