+++ platform/MCU/STM32F2xx/SPARK_Firmware_Driver/inc/dcd_flash_impl.h
@@ -19,69 +19,7 @@
 
 #include "dcd_flash.h"
 #include "dcd.h"
-#include "flash_mal.h"
-#include "string.h"
-
-/**
- * Implements access to the internal flash, providing the interface expected by dcd.h
- */
-class InternalFlashStore
-{
-public:
-    int eraseSector(unsigned address)
-    {
-        return !FLASH_EraseMemory(FLASH_INTERNAL, address, 1);
-    }
-
-    int write(const unsigned offset, const void* data, const unsigned size)
-    {
-        const uint8_t* data_ptr = (const uint8_t*)data;
-        const uint8_t* end_ptr  = data_ptr+size;
-        unsigned destination = offset;
-        FLASH_Unlock();
-        FLASH_ClearFlag(FLASH_FLAG_EOP | FLASH_FLAG_OPERR | FLASH_FLAG_WRPERR | FLASH_FLAG_PGAERR | FLASH_FLAG_PGPERR | FLASH_FLAG_PGSERR);
-
-        while (data_ptr < end_ptr)
-        {
-            FLASH_Status status;
-            const int max_tries = 10;
-            int tries = 0;
-
-            if ( !(destination & 0x03) && (end_ptr - data_ptr >= 4))  // have a whole word to write
-            {
-                while ((FLASH_COMPLETE != (status = FLASH_ProgramWord(destination, *(const uint32_t*)data_ptr))) && (tries++ < max_tries));
-                destination += 4;
-                data_ptr += 4;
-            }
-            else if ( !(destination & 0x01) && (end_ptr - data_ptr >= 2))  // have a half word to write
-            {
-                while ((FLASH_COMPLETE != (status = FLASH_ProgramHalfWord(destination, *(const uint32_t*)data_ptr))) && (tries++ < max_tries));
-                destination += 2;
-                data_ptr += 2;
-            }
-            else
-            {
-                while ((FLASH_COMPLETE != (status = FLASH_ProgramByte(destination, *data_ptr))) && (tries++ < max_tries));
-                destination++;
-                data_ptr++;
-            }
-
-        }
-        FLASH_Lock();
-        return (memcmp(dataAt(offset), data, size)) ? -1 : 0;
-    }
-
-    const uint8_t* dataAt(unsigned address)
-    {
-        return (const uint8_t*)address;
-    }
-
-    int read(unsigned offset, uint8_t* data, unsigned size)
-    {
-        memcpy(data, dataAt(offset), size);
-        return 0;
-    }
-};
+#include "flash_storage_impl.h"
 
 template <typename Store, unsigned sectorSize, unsigned DCD1, unsigned DCD2>
 class UpdateDCD : public DCD<Store, sectorSize, DCD1, DCD2>
